Break out tests into separate files
